move time processing to utterance
