+++ sktalk/corpus/parsing/parser.py
@@ -24,19 +24,6 @@ class InputFile(abc.ABC):
     def _extract_metadata(self):
         return {}
 
-    @staticmethod
-    def _to_timestamp(time_ms):
-        try:
-            time_ms = float(time_ms)
-        except ValueError:
-            return None
-        if time_ms > 86399999:
-            raise ValueError(f"timestamp {time_ms} exceeds 24h")
-        if time_ms < 0:
-            raise ValueError(f"timestamp {time_ms} negative")
-        time_dt = datetime.datetime.utcfromtimestamp(time_ms/1000)
-        return time_dt.strftime("%H:%M:%S.%f")[:-3]
-
     @classmethod
     def download(cls, url):              # noqa: W0613
         # download
