+++ sktalk/corpus/utterance.py
@@ -1,3 +1,4 @@
+import datetime
 import re
 from dataclasses import asdict
 from dataclasses import dataclass
@@ -9,7 +10,7 @@ from .participant import Participant
 class Utterance:
     utterance: str
     participant: Participant = None
-    time: str = None
+    time: list = None
     begin: str = None
     end: str = None
     metadata: dict[str, Any] = None
@@ -37,6 +38,9 @@ class Utterance:
         self.n_words = len(self.utterance_list)
         self.n_characters = sum(len(word) for word in self.utterance_list)
 
+        # calculate timestamps
+        self.begin, self.end = self._split_time(self.time)
+
     def get_audio(self):
         pass
 
@@ -46,6 +50,29 @@ class Utterance:
 
     def until(self, next_utt):
         return next_utt.time[0] - self.time[1]
+
+    def _split_time(self, time: list):
+        if time is None:
+            return None, None
+        begin, end = str(time).split(", ")
+        begin = begin.replace("(", "")
+        end = end.replace(")", "")
+        begin = self._to_timestamp(begin)
+        end = self._to_timestamp(end)
+        return (begin, end)
+
+    @staticmethod
+    def _to_timestamp(time_ms):
+        try:
+            time_ms = float(time_ms)
+        except ValueError:
+            return None
+        if time_ms > 86399999:
+            raise ValueError(f"timestamp {time_ms} exceeds 24h")
+        if time_ms < 0:
+            raise ValueError(f"timestamp {time_ms} negative")
+        time_dt = datetime.datetime.utcfromtimestamp(time_ms/1000)
+        return time_dt.strftime("%H:%M:%S.%f")[:-3]
 
     # TODO function: that prints summary of data, shows it to user
     # TODO function: create a pandas data frame with the utterances
