+++ tests/corpus/parsing/test_cha.py
@@ -7,25 +7,6 @@ from sktalk.corpus.parsing.cha import ChaFile
 from sktalk.corpus.parsing.parser import InputFile
 
 
-class TestParser:
-    milliseconds_timestamp = [
-        ["0", "00:00:00.000"],
-        ["1706326", "00:28:26.326"],
-        ["222222", "00:03:42.222"],
-        ["None", None]
-    ]
-
-    @pytest.mark.parametrize("milliseconds, timestamp", milliseconds_timestamp)
-    def test_to_timestamp(self, milliseconds, timestamp):
-        assert InputFile._to_timestamp(milliseconds) == timestamp   # noqa: W0212
-
-        with pytest.raises(ValueError, match="exceeds 24h"):
-            InputFile._to_timestamp("987654321")                    # noqa: W0212
-
-        with pytest.raises(ValueError, match="negative"):
-            InputFile._to_timestamp("-1")                           # noqa: W0212
-
-
 class TestChaFile:
     urls = [
         "https://ca.talkbank.org/data-orig/GCSAusE/01.cha",
@@ -61,15 +42,6 @@ class TestChaFile:
         language = parsed_cha.metadata["Languages"]
         assert language == ["eng"]
         # TODO assert that there are no empty utterances
-
-    def test_split_time(self):
-        time = "(1748070, 1751978)"
-        begin_end = ("00:29:08.070", "00:29:11.978")
-        assert ChaFile._split_time(time) == begin_end            # noqa: W0212
-
-        time = None
-        begin_end = (None, None)
-        assert ChaFile._split_time(time) == begin_end            # noqa: W0212
 
     unclean_clean = [
         [
