+++ tests/corpus/test_utterance.py
@@ -32,3 +32,32 @@ class TestUtterance():
     def test_until(self, convo_utts):
         utt1, utt2 = convo_utts
         assert utt1.until(utt2) == -100
+
+    milliseconds_timestamp = [
+        ["0", "00:00:00.000"],
+        ["1706326", "00:28:26.326"],
+        ["222222", "00:03:42.222"],
+        ["None", None]
+    ]
+
+    @pytest.mark.parametrize("milliseconds, timestamp", milliseconds_timestamp)
+    def test_to_timestamp(self, milliseconds, timestamp):
+        utt = Utterance(utterance="")
+        assert utt._to_timestamp(milliseconds) == timestamp   # noqa: W0212
+
+    def test_to_timestamp_errors(self):
+        utt = Utterance(utterance="")
+        with pytest.raises(ValueError, match="exceeds 24h"):
+            utt._to_timestamp("987654321")                    # noqa: W0212
+
+        with pytest.raises(ValueError, match="negative"):
+            Utterance._to_timestamp("-1")                           # noqa: W0212
+
+    time_begin_end = [[(1748070, 1751978), "00:29:08.070", "00:29:11.978"],
+                      [None, None, None]]
+
+    @pytest.mark.parametrize("time, begin, end", time_begin_end)
+    def test_split_time(self, time, begin, end):
+        utt = Utterance(utterance="", time=time)
+        assert utt.begin == begin
+        assert utt.end == end
